requirements.txt
streamlit>=1.35.0
openai>=1.30.0
openpyxl>=3.1.2
pandas>=2.2.0
numpy>=1.26.0
python-docx>=1.1.0
mammoth>=1.7.0
pdfminer.six>=20221105
xlrd>=2.0.1
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
